Major work on Sequencer --  o implemented Sequencer <-> MySQL database connection  o added test function for calling "backdoor" tests  o implemented threads for initializing connections to daemons  o some work on various other threads
tcs daemon
 o implemented delimiter-based read calls
 o added code to connect to TCS (real or emulated)

emulator
 o added power and tcs emulators
 o tcs emulator runs as daemon and gives realistic correctly-timed responses
 o cleanup slit emulator a bit

slit daemon
 o added is_open() and is_home() functions
 o implemented delimiter-based read calls

calib daemon
 o implemented delimiter-based read calls

power daemon
 o built up the skeleton a bit more

camera
 o added AstroCam::Interface::is_connected() function

miscellaneous
 o added TCS to run/ngps-config
 o removed special case for starting Java GUI TCS emulator in run/emulator
 o added is_home() function to PI interface
 o removed setting the NONBLOCK flag in Network::TcpSocket::Connect
   This affects all sockets; I think it's the right thing to do here.
 o added common/*_commands.h, a header file which defines const strings
   for all of the commands used by each module. I may change how this
   is done but for now I want their definitions to be defined outside
   the code -- Jennifer is going to need to know these.
 o added common/tcs_constants.h

dh

	modified:   CMakeLists.txt
	modified:   PI/pi.cpp
	modified:   calibd/CMakeLists.txt
	modified:   calibd/calib_interface.cpp
	modified:   calibd/calib_interface.h
	modified:   calibd/calibd.cpp
	modified:   calibd/calibd.h
	modified:   camerad/astrocam.cpp
	modified:   camerad/astrocam.h
	modified:   camerad/camerad.cpp
	new file:   common/calibd_commands.h
	new file:   common/camerad_commands.h
	new file:   common/filterd_commands.h
	new file:   common/flexured_commands.h
	new file:   common/focusd_commands.h
	new file:   common/powerd_commands.h
	modified:   common/sequencerd_commands.h
	modified:   common/slitd_commands.h
	new file:   common/tcs_constants.h
	new file:   common/tcsd_commands.h
	modified:   emulator/CMakeLists.txt
	new file:   emulator/emulatord_power.cpp
	new file:   emulator/emulatord_power.h
	modified:   emulator/emulatord_slit.cpp
	modified:   emulator/emulatord_slit.h
	new file:   emulator/emulatord_tcs.cpp
	new file:   emulator/emulatord_tcs.h
	new file:   emulator/power.cpp
	new file:   emulator/power.h
	modified:   emulator/slit.cpp
	new file:   emulator/tcs.cpp
	new file:   emulator/tcs.h
	modified:   flexured/flexured.cpp
	modified:   focusd/focusd.cpp
	modified:   powerd/CMakeLists.txt
	modified:   powerd/power_interface.cpp
	modified:   powerd/power_interface.h
	modified:   powerd/powerd.cpp
	modified:   powerd/powerd.h
	modified:   run/emulator
	modified:   run/ngps-config
	modified:   sequencerd/CMakeLists.txt
	modified:   sequencerd/sequence.cpp
	modified:   sequencerd/sequence.h
	modified:   sequencerd/sequencer_interface.cpp
	modified:   sequencerd/sequencer_interface.h
	modified:   sequencerd/sequencerd.cpp
	modified:   sequencerd/sequencerd.h
	modified:   slitd/slit_interface.cpp
	modified:   slitd/slit_interface.h
	modified:   slitd/slitd.cpp
	modified:   tcsd/CMakeLists.txt
	modified:   tcsd/tcs_interface.cpp
	modified:   tcsd/tcs_interface.h
	modified:   tcsd/tcsd.cpp
	modified:   tcsd/tcsd.h
	modified:   utils/network.cpp

+++ emulator/slit.cpp
@@ -141,7 +141,7 @@ namespace Slit {
       return( ERROR );
     }
 
-    if ( myaddr < 1 || myaddr > this->controller_info.size() ) {
+    if ( myaddr < 1 || myaddr > (long)this->controller_info.size() ) {
       std::cerr << function << "bad address " << myaddr << ". expected {0:" << this->controller_info.size() << "}\n";
       return( ERROR );
     }
